Separated routes.js into transcripts.routes.js and projects.routes.js. Note that the base URL for project APIs is "/api/v1/projects" !!! (It's updated in server.js)
