+++ back-end/server.js
@@ -1,6 +1,7 @@
-import express from "express"
-import cors from "cors"
-import transcripts from "./api/routes/routes.js"
+import express from "express";
+import cors from "cors";
+import transcripts from "./api/routes/transcripts.routes.js";
+import projects from "./api/routes/projects.routes.js";
 
 // This is the express config file
 const app = express()
@@ -9,6 +10,7 @@ app.use(cors())
 app.use(express.json())
 
 app.use("/api/v1/transcripts", transcripts)
+app.use("/api/v1/projects", projects)
 app.use("*", (req, res) => res.status(404).json({ error: "not found"}))
 
 export default app
